+++ .gitignore
@@ -16,3 +16,5 @@ dist-ssr
 **/.tsc-out/
 **/.turbo/
 **/tsconfig.tsbuildinfo
+.playwright-mcp/**/*
+.vite-hooks/**/*
