+++ apps/forum/package.json
@@ -4,25 +4,25 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "vinxi dev --port 3001",
-    "build": "vinxi build",
-    "start": "vinxi start"
+    "dev": "vite dev --port 3001",
+    "build": "vite build",
+    "start": "node .output/server/index.mjs"
   },
   "dependencies": {
     "@forum/api": "workspace:*",
     "@forum/db": "workspace:*",
     "@forum/ui": "workspace:*",
-    "@solidjs/router": "^0.16.1",
-    "@solidjs/start": "^1.3.2",
-    "better-auth": "^1.5.6",
-    "solid-js": "^1.9.12",
-    "vinxi": "^0.5.11"
+    "@tanstack/solid-router": "^1.168.19",
+    "@tanstack/solid-start": "^1.167.36",
+    "better-auth": "^1.6.3",
+    "solid-js": "^1.9.12"
   },
   "devDependencies": {
     "@forum/config": "workspace:*",
     "@tailwindcss/vite": "^4.2.2",
-    "tailwindcss": "^4.2.2",
     "daisyui": "^5.5.19",
-    "vite": "^8.0.3"
+    "tailwindcss": "^4.2.2",
+    "vite": "^7.0.0",
+    "vite-plugin-solid": "^2.11.0"
   }
 }
