--- apps/forum/src/entry-client.tsx
@@ -1,4 +1,0 @@
-import { mount, StartClient } from "@solidjs/start/client";
-
-// biome-ignore lint/style/noNonNullAssertion: SolidStart convention — #app always exists
-mount(() => <StartClient />, document.getElementById("app")!);
