+++ apps/forum/src/routeTree.gen.ts
@@ -1,0 +1,177 @@
+/* eslint-disable */
+
+// @ts-nocheck
+
+// noinspection JSUnusedGlobalSymbols
+
+// This file was automatically generated by TanStack Router.
+// You should NOT make any changes in this file as it will be overwritten.
+// Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
+
+import { Route as rootRouteImport } from './routes/__root'
+import { Route as IndexRouteImport } from './routes/index'
+import { Route as CategoryIndexRouteImport } from './routes/$category/index'
+import { Route as AuthSignUpRouteImport } from './routes/auth/sign-up'
+import { Route as AuthSignInRouteImport } from './routes/auth/sign-in'
+import { Route as CategorySubIndexRouteImport } from './routes/$category/$sub/index'
+import { Route as CategorySubTopicIndexRouteImport } from './routes/$category/$sub/$topic/index'
+
+const IndexRoute = IndexRouteImport.update({
+  id: '/',
+  path: '/',
+  getParentRoute: () => rootRouteImport,
+} as any)
+const CategoryIndexRoute = CategoryIndexRouteImport.update({
+  id: '/$category/',
+  path: '/$category/',
+  getParentRoute: () => rootRouteImport,
+} as any)
+const AuthSignUpRoute = AuthSignUpRouteImport.update({
+  id: '/auth/sign-up',
+  path: '/auth/sign-up',
+  getParentRoute: () => rootRouteImport,
+} as any)
+const AuthSignInRoute = AuthSignInRouteImport.update({
+  id: '/auth/sign-in',
+  path: '/auth/sign-in',
+  getParentRoute: () => rootRouteImport,
+} as any)
+const CategorySubIndexRoute = CategorySubIndexRouteImport.update({
+  id: '/$category/$sub/',
+  path: '/$category/$sub/',
+  getParentRoute: () => rootRouteImport,
+} as any)
+const CategorySubTopicIndexRoute = CategorySubTopicIndexRouteImport.update({
+  id: '/$category/$sub/$topic/',
+  path: '/$category/$sub/$topic/',
+  getParentRoute: () => rootRouteImport,
+} as any)
+
+export interface FileRoutesByFullPath {
+  '/': typeof IndexRoute
+  '/auth/sign-in': typeof AuthSignInRoute
+  '/auth/sign-up': typeof AuthSignUpRoute
+  '/$category/': typeof CategoryIndexRoute
+  '/$category/$sub/': typeof CategorySubIndexRoute
+  '/$category/$sub/$topic/': typeof CategorySubTopicIndexRoute
+}
+export interface FileRoutesByTo {
+  '/': typeof IndexRoute
+  '/auth/sign-in': typeof AuthSignInRoute
+  '/auth/sign-up': typeof AuthSignUpRoute
+  '/$category': typeof CategoryIndexRoute
+  '/$category/$sub': typeof CategorySubIndexRoute
+  '/$category/$sub/$topic': typeof CategorySubTopicIndexRoute
+}
+export interface FileRoutesById {
+  __root__: typeof rootRouteImport
+  '/': typeof IndexRoute
+  '/auth/sign-in': typeof AuthSignInRoute
+  '/auth/sign-up': typeof AuthSignUpRoute
+  '/$category/': typeof CategoryIndexRoute
+  '/$category/$sub/': typeof CategorySubIndexRoute
+  '/$category/$sub/$topic/': typeof CategorySubTopicIndexRoute
+}
+export interface FileRouteTypes {
+  fileRoutesByFullPath: FileRoutesByFullPath
+  fullPaths:
+    | '/'
+    | '/auth/sign-in'
+    | '/auth/sign-up'
+    | '/$category/'
+    | '/$category/$sub/'
+    | '/$category/$sub/$topic/'
+  fileRoutesByTo: FileRoutesByTo
+  to:
+    | '/'
+    | '/auth/sign-in'
+    | '/auth/sign-up'
+    | '/$category'
+    | '/$category/$sub'
+    | '/$category/$sub/$topic'
+  id:
+    | '__root__'
+    | '/'
+    | '/auth/sign-in'
+    | '/auth/sign-up'
+    | '/$category/'
+    | '/$category/$sub/'
+    | '/$category/$sub/$topic/'
+  fileRoutesById: FileRoutesById
+}
+export interface RootRouteChildren {
+  IndexRoute: typeof IndexRoute
+  AuthSignInRoute: typeof AuthSignInRoute
+  AuthSignUpRoute: typeof AuthSignUpRoute
+  CategoryIndexRoute: typeof CategoryIndexRoute
+  CategorySubIndexRoute: typeof CategorySubIndexRoute
+  CategorySubTopicIndexRoute: typeof CategorySubTopicIndexRoute
+}
+
+declare module '@tanstack/solid-router' {
+  interface FileRoutesByPath {
+    '/': {
+      id: '/'
+      path: '/'
+      fullPath: '/'
+      preLoaderRoute: typeof IndexRouteImport
+      parentRoute: typeof rootRouteImport
+    }
+    '/$category/': {
+      id: '/$category/'
+      path: '/$category'
+      fullPath: '/$category/'
+      preLoaderRoute: typeof CategoryIndexRouteImport
+      parentRoute: typeof rootRouteImport
+    }
+    '/auth/sign-up': {
+      id: '/auth/sign-up'
+      path: '/auth/sign-up'
+      fullPath: '/auth/sign-up'
+      preLoaderRoute: typeof AuthSignUpRouteImport
+      parentRoute: typeof rootRouteImport
+    }
+    '/auth/sign-in': {
+      id: '/auth/sign-in'
+      path: '/auth/sign-in'
+      fullPath: '/auth/sign-in'
+      preLoaderRoute: typeof AuthSignInRouteImport
+      parentRoute: typeof rootRouteImport
+    }
+    '/$category/$sub/': {
+      id: '/$category/$sub/'
+      path: '/$category/$sub'
+      fullPath: '/$category/$sub/'
+      preLoaderRoute: typeof CategorySubIndexRouteImport
+      parentRoute: typeof rootRouteImport
+    }
+    '/$category/$sub/$topic/': {
+      id: '/$category/$sub/$topic/'
+      path: '/$category/$sub/$topic'
+      fullPath: '/$category/$sub/$topic/'
+      preLoaderRoute: typeof CategorySubTopicIndexRouteImport
+      parentRoute: typeof rootRouteImport
+    }
+  }
+}
+
+const rootRouteChildren: RootRouteChildren = {
+  IndexRoute: IndexRoute,
+  AuthSignInRoute: AuthSignInRoute,
+  AuthSignUpRoute: AuthSignUpRoute,
+  CategoryIndexRoute: CategoryIndexRoute,
+  CategorySubIndexRoute: CategorySubIndexRoute,
+  CategorySubTopicIndexRoute: CategorySubTopicIndexRoute,
+}
+export const routeTree = rootRouteImport
+  ._addFileChildren(rootRouteChildren)
+  ._addFileTypes<FileRouteTypes>()
+
+import type { getRouter } from './router.tsx'
+import type { createStart } from '@tanstack/solid-start'
+declare module '@tanstack/solid-start' {
+  interface Register {
+    ssr: true
+    router: Awaited<ReturnType<typeof getRouter>>
+  }
+}
