+++ apps/forum/src/routes/index.tsx
@@ -1,5 +1,6 @@
-import { A } from "@solidjs/router";
-import { createResource, For, Show, Suspense } from "solid-js";
+import { createFileRoute, Link } from "@tanstack/solid-router";
+import { createServerFn } from "@tanstack/solid-start";
+import { For, Show } from "solid-js";
 import { apiFetch } from "@/lib/api";
 
 interface Subcategory {
@@ -18,56 +19,58 @@ interface Category {
   subcategories: Subcategory[];
 }
 
-export default function Home() {
-  const [categories] = createResource(() =>
-    apiFetch<Category[]>("/categories"),
-  );
+const fetchCategories = createServerFn({ method: "GET" }).handler(async () => {
+  return await apiFetch<Category[]>("/categories");
+});
+
+export const Route = createFileRoute("/")({
+  loader: () => fetchCategories(),
+  component: Home,
+});
+
+function Home() {
+  const categories = Route.useLoaderData();
 
   return (
     <div>
-      <h1 class="text-2xl font-bold mb-6">Forum</h1>
-      <Suspense fallback={<span class="loading loading-spinner loading-lg" />}>
-        <Show
-          when={categories()}
-          fallback={<p class="text-base-content/60">No categories yet.</p>}
-        >
-          {(cats) => (
-            <div class="space-y-4">
-              <For each={cats()}>
-                {(category) => (
-                  <div class="card bg-base-100 shadow">
-                    <div class="card-body">
-                      <h2 class="card-title">
-                        <Show when={category.icon}>
-                          <span>{category.icon}</span>
-                        </Show>
-                        {category.name}
-                      </h2>
-                      <Show when={category.description}>
-                        <p class="text-base-content/60">
-                          {category.description}
-                        </p>
-                      </Show>
-                      <div class="mt-2">
-                        <For each={category.subcategories}>
-                          {(sub) => (
-                            <A
-                              href={`/${category.slug}/${sub.slug}`}
-                              class="btn btn-ghost btn-sm"
-                            >
-                              {sub.name}
-                            </A>
-                          )}
-                        </For>
-                      </div>
-                    </div>
+      <h1 class="mb-6 text-2xl font-bold">Forum</h1>
+      <Show
+        when={(categories()?.length ?? 0) > 0}
+        fallback={<p class="text-base-content/60">No categories yet.</p>}
+      >
+        <div class="space-y-4">
+          <For each={categories()}>
+            {(category) => (
+              <div class="card bg-base-100 shadow">
+                <div class="card-body">
+                  <h2 class="card-title">
+                    <Show when={category.icon}>
+                      <span>{category.icon}</span>
+                    </Show>
+                    {category.name}
+                  </h2>
+                  <Show when={category.description}>
+                    <p class="text-base-content/60">{category.description}</p>
+                  </Show>
+                  <div class="mt-2">
+                    <For each={category.subcategories}>
+                      {(sub) => (
+                        <Link
+                          to="/$category/$sub"
+                          params={{ category: category.slug, sub: sub.slug }}
+                          class="btn btn-ghost btn-sm"
+                        >
+                          {sub.name}
+                        </Link>
+                      )}
+                    </For>
                   </div>
-                )}
-              </For>
-            </div>
-          )}
-        </Show>
-      </Suspense>
+                </div>
+              </div>
+            )}
+          </For>
+        </div>
+      </Show>
     </div>
   );
 }
