+++ packages/api/package.json
@@ -13,13 +13,13 @@
   },
   "dependencies": {
     "@forum/db": "workspace:*",
-    "better-auth": "^1.5.6",
-    "drizzle-orm": "^0.45.1",
-    "hono": "^4.12.9",
-    "postgres": "^3.4.8"
+    "better-auth": "^1.6.3",
+    "drizzle-orm": "^0.45.2",
+    "hono": "^4.12.14",
+    "postgres": "^3.4.9"
   },
   "devDependencies": {
     "@forum/config": "workspace:*",
-    "@types/bun": "^1.3.11"
+    "@types/bun": "^1.3.12"
   }
 }
