+++ packages/db/package.json
@@ -15,12 +15,12 @@
     "db:seed": "bun --bun --env-file ../../.env run scripts/seed.ts"
   },
   "dependencies": {
-    "drizzle-orm": "^0.45.1",
-    "postgres": "^3.4.8"
+    "drizzle-orm": "^0.45.2",
+    "postgres": "^3.4.9"
   },
   "devDependencies": {
     "@forum/config": "workspace:*",
     "drizzle-kit": "^0.31.10",
-    "@types/bun": "^1.3.11"
+    "@types/bun": "^1.3.12"
   }
 }
