+++ packages/ui/src/components/Modal.tsx
@@ -1,5 +1,4 @@
-import { type ParentProps, Show } from "solid-js";
-import { Portal } from "solid-js/web";
+import { createEffect, type ParentProps } from "solid-js";
 
 interface ModalProps {
   open: boolean;
@@ -9,41 +8,40 @@ interface ModalProps {
 }
 
 export function Modal(props: ParentProps<ModalProps>) {
+  let dialogRef!: HTMLDialogElement;
+
+  createEffect(() => {
+    if (props.open) {
+      dialogRef?.showModal();
+    } else {
+      dialogRef?.close();
+    }
+  });
+
   return (
-    <Show when={props.open}>
-      <Portal>
-        <div class="fixed inset-0 z-50 flex items-center justify-center">
-          {/* Backdrop — button for a11y, closes on click */}
-          <button
-            class="absolute inset-0 h-full w-full cursor-default border-0 bg-black/40 p-0"
-            onClick={props.onClose}
-            aria-label="Close dialog"
-            tabIndex={-1}
-          />
-          {/* Dialog box — relative z-index sits above the backdrop */}
-          <div
-            class={`modal-box relative z-10 rounded bg-base-100 p-6 shadow ${props.class ?? ""}`}
-            role="dialog"
-            aria-modal="true"
-            aria-labelledby={props.title ? "modal-title" : undefined}
-          >
-            <Show when={props.title}>
-              <div class="mb-4 flex items-center justify-between">
-                <h3 id="modal-title" class="text-lg font-bold">
-                  {props.title}
-                </h3>
-                <button
-                  class="btn btn-circle btn-ghost btn-sm"
-                  onClick={props.onClose}
-                >
-                  ✕
-                </button>
-              </div>
-            </Show>
-            {props.children}
+    <dialog
+      ref={dialogRef}
+      class="modal"
+      onClose={props.onClose}
+      onClick={(e) => {
+        // Close when clicking the backdrop (outside modal-box)
+        if (e.target === dialogRef) props.onClose();
+      }}
+    >
+      <div class={`modal-box ${props.class ?? ""}`}>
+        {props.title && (
+          <div class="mb-4 flex items-center justify-between">
+            <h3 class="text-lg font-bold">{props.title}</h3>
+            <button
+              class="btn btn-circle btn-ghost btn-sm"
+              onClick={props.onClose}
+            >
+              ✕
+            </button>
           </div>
-        </div>
-      </Portal>
-    </Show>
+        )}
+        {props.children}
+      </div>
+    </dialog>
   );
 }
